added separate props to styled
