+++ src/components/Dashboard/Content/NewOrders/NewOrders.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { styled } from "@mui/system";
 import notifaction from "src/images/svg/notifaction.svg";
-import { ViewButton } from "src/components/UI/commonStyles";
+import { ViewAll, ViewButton } from "src/components/UI/commonStyles";
 
 const BlockNewOrders = styled("div")({
   width: "15%",
@@ -85,12 +85,6 @@ const NewOrdersDescrSpan = styled("span")({
   lineHeight: "13px",
   textAlign: "center",
   color: "#7E89A8",
-});
-
-export const ViewAll = styled("div")({
-  display: "flex",
-  justifyContent: "center",
-  marginTop: "30px",
 });
 
 const NewOrders = () => {
